+++ packages/terra-button/src/terra-dev-site/doc/example/ButtonVariant.jsx
@@ -9,6 +9,7 @@ const ButtonVariant = () => (
   <div>
     <Button text="Neutral" style={buttonStyle} />
     <Button text="Emphasis" variant="emphasis" style={buttonStyle} />
+    <Button text="Ghost" variant="ghost" style={buttonStyle} />
     <Button text="De-emphasis" variant="de-emphasis" style={buttonStyle} />
     <Button text="Action" variant="action" icon={<IconEdit />} style={buttonStyle} />
     <Button text="Utility" variant="utility" icon={<IconEdit />} style={buttonStyle} />