+++ packages/terra-doc-template/src/ExampleTemplate.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import SyntaxHighlighter, { registerLanguage } from 'react-syntax-highlighter/prism-light';
-import jsx from 'react-syntax-highlighter/languages/prism/jsx';
+import SyntaxHighlighter, { registerLanguage } from 'react-syntax-highlighter/dist/prism-light';
+import jsx from 'react-syntax-highlighter/dist/languages/prism/jsx';
 import classNames from 'classnames/bind';
 import styles from './ExampleTemplate.module.scss';
 import theme from './syntax-highlighter-theme';