+++ packages/terra-arrange/CHANGELOG.md
@@ -4,6 +4,12 @@ Changelog
 Unreleased
 ----------
 
+3.9.0 - (May 1, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+* Restructured arrange examples to better demonstrate alignment functionality
+
 3.8.0 - (April 16, 2019)
 ------------------
 ### Changed