+++ packages/terra-form-radio/CHANGELOG.md
@@ -4,6 +4,14 @@ ChangeLog
 Unreleased
 ----------
 
+3.12.0 - (May 1, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+### Added
+* Doc site examples that can be toggled to demonstrate isInline and isInvalid props
+
 3.11.0 - (April 16, 2019)
 ------------------
 ### Changed