+++ packages/terra-profile-image/CHANGELOG.md
@@ -4,6 +4,12 @@ Changelog
 Unreleased
 ----------
 
+3.8.0 - (May 1, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+* Updated WDIO tests to use themeCombinationOfCustomProperties instead of themeEachCustomProperty
+
 3.7.0 - (April 16, 2019)
 ------------------
 ### Changed