+++ packages/terra-responsive-element/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+4.7.0 - (May 1, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 4.6.0 - (April 16, 2019)
 ------------------
 ### Changed